Refactor : CarFactory.java의 패키지 (utils -> domain) 이동
